+++ src/main/java/com/lowdragmc/mbd2/core/mixins/ChunkMixin.java
@@ -7,7 +7,6 @@ import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.chunk.LevelChunk;
-import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -20,7 +19,10 @@ public class ChunkMixin {
     @Final @Shadow Level level;
 
     // We want to be as quick as possible here
-    @Inject(method = "setBlockState", at =@At(value = "INVOKE", opcode = Opcodes.GETFIELD, target = "Lnet/minecraft/world/level/block/state/BlockState;onPlace(Lnet/minecraft/world/level/Level;Lnet/minecraft/core/BlockPos;Lnet/minecraft/world/level/block/state/BlockState;Z)V"))
+    @Inject(method = "setBlockState", at =@At(
+            value = "INVOKE",
+            target = "Lnet/minecraft/world/level/block/state/BlockState;hasBlockEntity()Z",
+            ordinal = 2))
     private void onAddingBlock(BlockPos pos, BlockState state, boolean isMoving, CallbackInfoReturnable<BlockState> cir) {
         MinecraftServer server = level.getServer();
         if (server != null) {